requirements.txt
Django==4.2.3
djangorestframework==3.14.0
psycopg2==2.9.6
pytest==7.4.0
pytest-django==4.5.2
pytz==2023.3
setuptools==59.6.0